moving test files into  folder
